requirements.txt
# Core dependencies
crewai>=0.175.0
langchain-ollama>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
watchdog>=3.0.0
flask>=3.0.0
streamlit>=1.29.0
pandas>=2.1.0
numpy>=1.24.0
rich>=13.7.0
texttable>=1.6.7
termgraph>=0.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
typer>=0.9.0

# Optional features
scikit-learn>=1.3.0
scipy>=1.11.0
